requirements.txt
paho-mqtt==2.1.0
pytest==8.0.0